fvfvf

+++ app/src/main/java/com/multimedia/aes/gestnet_nucleo/BBDD/GuardarDisposiciones.java
@@ -76,7 +76,7 @@ public class GuardarDisposiciones extends AsyncTask<Void,Void,Void> {
 
         int id_disposicion_servicio;
         String nombre_disposicion;
-        int coste_disposicion, precio_disposicion;
+        double coste_disposicion, precio_disposicion;
         boolean esta = false;
         JSONObject jsonObject = new JSONObject(json);
         JSONArray jsonArray = jsonObject.getJSONArray("disposiciones_servicio");
@@ -110,13 +110,13 @@ public class GuardarDisposiciones extends AsyncTask<Void,Void,Void> {
             if (jsonArray.getJSONObject(i).getString("coste_disposicion").equals("null") || jsonArray.getJSONObject(i).getString("coste_disposicion").equals("")) {
                 coste_disposicion = 0;
             } else {
-                coste_disposicion = jsonArray.getJSONObject(i).getInt("coste_disposicion");
+                coste_disposicion = jsonArray.getJSONObject(i).getDouble("coste_disposicion");
             }
 
             if (jsonArray.getJSONObject(i).getString("precio_disposicion").equals("null") || jsonArray.getJSONObject(i).getString("precio_disposicion").equals("")) {
                 precio_disposicion = 0;
             } else {
-                precio_disposicion = jsonArray.getJSONObject(i).getInt("precio_disposicion");
+                precio_disposicion = jsonArray.getJSONObject(i).getDouble("precio_disposicion");
             }
             if (!esta) {
                 if (DisposicionesDAO.newDisposiciones(context, id_disposicion_servicio, nombre_disposicion, coste_disposicion, precio_disposicion)) {
